+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: oozsertt <oozsertt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/15 20:14:32 by user42            #+#    #+#             */
-/*   Updated: 2020/10/12 19:57:03 by oozsertt         ###   ########.fr       */
+/*   Updated: 2021/10/05 14:58:25 by oozsertt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const	char *s)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
