+++ ft_utoa_base_uppercase.c
@@ -6,7 +6,7 @@
 /*   By: oozsertt <oozsertt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/30 16:26:24 by oozsertt          #+#    #+#             */
-/*   Updated: 2021/07/15 10:50:44 by oozsertt         ###   ########.fr       */
+/*   Updated: 2021/10/05 15:04:23 by oozsertt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ static int	ft_get_size_nbr(long nbr, int base_len)
 	return (i);
 }
 
-static char		place_in_base(long nbr, int len_base)
+static char	place_in_base(long nbr, int len_base)
 {
 	if (nbr % len_base >= 10)
 		return ((nbr % len_base) - 10 + 'A');
@@ -33,11 +33,11 @@ static char		place_in_base(long nbr, int len_base)
 		return ((nbr % len_base) + '0');
 }
 
-char		*ft_utoa_base_uppercase(unsigned int n, char *base)
+char	*ft_utoa_base_uppercase(unsigned int n, char *base)
 {
 	char	*str;
-	int		len_str;
-	int		len_base;
+	size_t	len_str;
+	size_t	len_base;
 	long	nbr;
 
 	len_base = ft_strlen(base);
@@ -47,7 +47,7 @@ char		*ft_utoa_base_uppercase(unsigned int n, char *base)
 		return (NULL);
 	nbr = n;
 	len_str = ft_get_size_nbr(nbr, len_base);
-	str = (char*)malloc(sizeof(char) * (len_str + 1));
+	str = (char *)malloc(sizeof(char) * (len_str + 1));
 	if (str == NULL)
 		return (NULL);
 	str[len_str] = '\0';
